requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
annotated-types==0.7.0
attrs==25.3.0
binance-futures-connector==4.1.0
blinker==1.9.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
dateparser==1.2.1
Flask==3.1.0
flask-cors==5.0.1
frozenlist==1.6.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
multidict==6.4.3
numpy==2.2.5
pandas==2.2.3
pandas_ta==0.3.14b0
propcache==0.3.1
pycryptodome==3.22.0
pydantic==2.11.3
pydantic_core==2.33.1
python-binance==1.0.28
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
regex==2024.11.6
requests==2.32.3
setuptools==79.0.0
six==1.17.0
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.4.0
websocket-client==1.8.0
websockets==15.0.1
Werkzeug==3.1.3
yarl==1.20.0
gunicorn
uvicorn
psycopg2-binary